wikistream-enricher/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
kafka-python==2.1.2
requests==2.32.3
urllib3==2.3.0
